JSON save output
-JSON Output Option:
At the end of the scan, the program asks the user if they want to save the results as a JSON file.
If the user agrees, results are saved in the same directory as the script, named scan_results.json.

-Readable File Path: Outputs the absolute path of the JSON file after saving.

-Results Format:
Hosts are saved along with their IP addresses.
Each host includes a list of open ports and detected services.